finalize templates and fix specs
