+++ src/Witnesses/SystemInfo/Witness.ts
@@ -1,10 +1,9 @@
 import { XyoWitness, XyoWitnessConfig } from '../../core'
 import { XyoSystemInfoPayload } from './Payload'
-import { systemInfoTemplate } from './Template'
 
 export class XyoSystemInfoWitness<T extends XyoSystemInfoPayload = XyoSystemInfoPayload, C extends XyoWitnessConfig<T> = XyoWitnessConfig<T>> extends XyoWitness<T, C> {
   // TODO: Handle different template types for various system witnesses
-  constructor(config: C = { schema: 'network.xyo.system.info', template: systemInfoTemplate } as C) {
+  constructor(config: C = { schema: 'network.xyo.system.info' } as C) {
     super({ ...config })
   }
