+++ src/browser/Witnesses/SystemInfo/Template.ts
@@ -1,14 +1,11 @@
 import { Parser } from 'bowser'
 
-import { XyoSystemInfoPayload } from './Payload'
-
-export const systemInfoTemplate: XyoSystemInfoPayload = {
+export const SystemInfoBrowserWitnessTemplate = () => ({
   bowser: {
     browser: {} as Parser.ParsedResult['browser'],
     engine: {} as Parser.ParsedResult['engine'],
     os: {} as Parser.ParsedResult['os'],
     platform: {} as Parser.ParsedResult['platform'],
   },
-  schema: 'network.xyo.system.info',
-  systeminformation: {},
-}
+  schema: 'network.xyo.system.info.browser',
+})
