+++ src/node/Witnesses/SystemInfo/Witness.ts
@@ -4,24 +4,7 @@ import { get } from 'systeminformation'
 import { XyoWitnessConfig } from '../../../core'
 import { XyoSystemInfoWitness } from '../../../Witnesses'
 import { XyoSystemInfoNodePayload } from './Payload'
-
-const defaultSystemInfoConfig = () => {
-  return {
-    audio: '*',
-    battery: '*',
-    bluetooth: '*',
-    cpu: '*',
-    diskLayout: '*',
-    graphics: '*',
-    mem: '*',
-    networkInterfaces: '*',
-    osInfo: '*',
-    printer: '*',
-    system: '*',
-    usb: '*',
-    wifiInterfaces: '*',
-  }
-}
+import { defaultSystemInfoConfig, systemInfoNodeWitnessTemplate } from './Template'
 
 export interface XyoSystemInfoNodeWitnessConfig<T extends XyoSystemInfoNodePayload = XyoSystemInfoNodePayload> extends XyoWitnessConfig<T> {
   nodeValues?: Record<string, string>
@@ -31,13 +14,14 @@ export class XyoSystemInfoNodeWitness<
   T extends XyoSystemInfoNodePayload = XyoSystemInfoNodePayload,
   C extends XyoSystemInfoNodeWitnessConfig<T> = XyoSystemInfoNodeWitnessConfig<T>
 > extends XyoSystemInfoWitness<T, C> {
-  constructor(config: C = { schema: XyoSystemInfoNodeWitness.schema } as C, baseSchema = XyoSystemInfoNodeWitness.schema) {
-    super(config, baseSchema)
+  constructor(config: C = { schema: 'network.xyo.system.info.node' } as C) {
+    super({
+      ...{ template: systemInfoNodeWitnessTemplate() },
+      ...config,
+    })
   }
   override async observe(fields?: Partial<T>) {
     const node = await get(this.config.nodeValues ?? defaultSystemInfoConfig())
     return await super.observe(merge({ node }, fields))
   }
-
-  public static schema = 'network.xyo.system.info.node'
 }
